Fixed a bug in 4th order scalar diffusion and scalar flux probes.
The error in 4th order diffusion was in the calculation of the fluxes
at solid boundaries due accessing Fl(i,j,k) instead of Fl(xi,yj,zk).
The error in scalar flux probes was caused by accessing the array out of
bounds when the probes_flux was empty. Now optional arguments are used,
but the values are only computed when averaging is on.
Also, the diffusivity in AddScalarDiffVector had to be divided by 2.